requirements.txt
ShopifyAPI
python-dotenv
requests
pandas
numpy